gpopa

+++ convert_base.c
@@ -56,7 +56,7 @@ char *citoa(long int num, char *str, int base, int prec, int *size)
  * @size: size of str
  * Return: a character array.
  */
-char *cuitoa(unsigned int num, char *str, int base, int prec, int *size)
+char *cuitoa(unsigned long int num, char *str, int base, int prec, int *size)
 {
 	int i = 0;
 	bool isNegative = false;
